requirements.txt
discord
dateparser
beautifulsoup4
pytz
requests
greenlet
wolframalpha
PyDictionary
pyimgur
wolframalpha
pillow
googletrans
